+++ packages/digit_data_model/lib/models/entities/project.dart
@@ -184,8 +184,10 @@ class ProjectAdditionalFields extends AdditionalFields
 @MappableClass(ignoreNull: true, discriminatorValue: MappableClass.useAsDefault)
 class ProjectAdditionalDetails with ProjectAdditionalDetailsMappable {
   final ProjectTypeModel? projectType;
+  final ProjectTypeModel? additionalProjectType;
 
   ProjectAdditionalDetails({
     this.projectType,
+    this.additionalProjectType,
   }) : super();
 }